+++ README.md
@@ -1,10 +1,1 @@
 # Big Data
-
-# Unit 1
-* [PRACTICE #1](https://github.com/ArturoCeron/BigData/blob/Unidad_1/Practices/Practice_1/README.md)
-* [PRACTICE #2](https://github.com/ArturoCeron/BigData/blob/Unidad_1/Practices/Practice_2/README.md)
-* [PRACTICE #3](https://github.com/ArturoCeron/BigData/blob/Unidad_1/Practices/Practice_3/README.md)
-* [PRACTICE #4](https://github.com/ArturoCeron/BigData/blob/Unidad_1/Practices/Practice_4/README.md)
-* [PRACTICE #5](https://github.com/ArturoCeron/BigData/blob/Unidad_1/Practices/Practice_5/README.md)
-* [PAIR CODING](https://github.com/ArturoCeron/BigData/blob/Unidad_1/Investigation/README.md)
-* [PEARSON CORRELATION COEFFICIENT](https://github.com/ArturoCeron/BigData/blob/Unidad_1/Homeworks/README.md)
