+++ Unit_1/Practices/Practice_3/Practica3.scala
