+++ Unit_1/Practices/Practice_5/Practica5.scala
